Move CopyActionFlags and SelectionFlags to top-level NameSpace
- Potentially breaking change if anyone was using it. This will make it easier to use though.
